Add files via upload
A really simple tool to find which gene your mitochondrial base position is in. It is an Rshiny tool, so you can just run the code and change the path links as needed. mtDNA01 is a redundant piece of data for now, but I have plans to improve this tool to include options for visualization.