Add files via upload
This is a very simple (for now) Rshiny tool to find which gene (or genes) your mitochondrial base position is located in. Currently, it just takes in an input to do that and outputs the name of the genes. However, I have plans to make it track and visualise variants. I made this out of the simple issue of having to search for the coordinate of genes every single time I wanted to do something with the mitochondrial genome. Hope this is useful!

The code itself is very simple, just replace the paths of the files with wherever they are located on your system.